+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Multiplicative method, PredictMultiplicative.
+### Changed
+- Predict changed to PredictAdditive to allow for Multiplcative predition.
 
 ## [v0.1.0] - 2019-12-15
 ### Added
 - Predict function to use Holt-Winters exponential smoothing to predict future values and smooth existing values.
 
+[Unreleased]: https://github.com/jthomperoo/holtwinters/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/jthomperoo/holtwinters/releases/tag/v0.1.0